@@ -5,13 +5,15 @@
   "description": "Tool to detect idle users inside a WhatsApp community",
   "author": "Pez Cuckow <email@pezcuckow.com>",
   "main": "index.js",
+  "type": "module",
   "dependencies": {
     "debug-level": "^3.0.0",
+    "inquirer": "^9.2.11",
     "qrcode-terminal": "^0.12.0",
-    "whatsapp-web.js": "git@github.com:pedroslopez/whatsapp-web.js.git#152f0a64cdffdfb4d638043f537f87b2a6ebc492"
+    "whatsapp-web.js": "git@github.com:pedroslopez/whatsapp-web.js.git#ecf19eb9a59fd9bc24457694a78716ea892cde91"
   },
   "devDependencies": {
-    "eslint": "^8.49.0",
+    "eslint": "^8.50.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.28.1",
     "prettier": "^3.0.3"
